hide CodeBlock
